requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --universal --refresh --output-file requirements.txt
amadeus==12.0.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
click==8.1.8
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
fastapi==0.115.12
    # via -r requirements.in
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.1.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.39
    # via -r requirements.in
starlette==0.46.1
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.34.0
    # via -r requirements.in
